#121 | refactor: Self Role Application Operations Have Been Separated with `RoleSelfApplicationController` and Permissions Have Been Renamed
